pyproject.toml
[project]
name = "autoresearch"
version = "0.2.0"
description = "Autonomous pretraining research — CPU lightweight edition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.7.0",
    "torch>=2.2.0",
    "anthropic>=0.30.0",
    "python-dotenv>=1.2.2",
]
